+++ lib/braintree/transaction.rb
@@ -25,6 +25,7 @@ module Braintree
       CVV          = "cvv"
       Duplicate    = "duplicate"
       Fraud        = "fraud"
+      RiskThreshold = "risk_threshold"
       ThreeDSecure = "three_d_secure"
       TokenIssuance = "token_issuance"
       Unrecognized = "unrecognized"