+++ spec/integration/braintree/transaction_spec.rb
@@ -909,6 +909,32 @@ describe Braintree::Transaction do
         end
       end
 
+      it "exposes the risk_threshold gateway rejection reason (via test cc num)" do
+        with_advanced_fraud_integration_merchant do
+          result = Braintree::Transaction.sale(
+            :amount => Braintree::Test::TransactionAmounts::Authorize,
+            :credit_card => {
+              :number => Braintree::Test::CreditCardNumbers::RiskThreshold,
+              :expiration_date => "05/2017",
+              :cvv => "333"
+            }
+          )
+          result.success?.should == false
+          result.transaction.gateway_rejection_reason.should == Braintree::Transaction::GatewayRejectionReason::RiskThreshold
+        end
+      end
+
+      it "exposes the risk_threshold gateway rejection reason (via test test nonce)" do
+        with_advanced_fraud_integration_merchant do
+          result = Braintree::Transaction.sale(
+            :amount => Braintree::Test::TransactionAmounts::Authorize,
+            :payment_method_nonce => Braintree::Test::Nonce::GatewayRejectedRiskThresholds,
+          )
+          result.success?.should == false
+          result.transaction.gateway_rejection_reason.should == Braintree::Transaction::GatewayRejectionReason::RiskThreshold
+        end
+      end
+
       it "exposes the token issuance gateway rejection reason" do
         result = Braintree::Transaction.sale(
           :amount => Braintree::Test::TransactionAmounts::Authorize,